requirements.txt
#torch==2.3.1+cu121 torchvision==0.18.1+cu121 -f https://download.pytorch.org/whl/torch_stable.html
torchsummary
ipykernel
numpy
pandas
matplotlib
python-dotenv
scikit-learn
boto3
tqdm